+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Adding support for timeout fields in the App CR.
+
 ## [6.14.0] - 2022-09-13
 
 ### Changed